+++ kinto/core/utils.py
@@ -226,6 +226,8 @@ class COMPARISON(Enum):
     EXCLUDE = 'exclude'
     LIKE = 'like'
     HAS = 'has'
+    CONTAINS = 'contains'
+    CONTAINS_ANY = 'contains_any'
 
 
 def reapply_cors(request, response):